+++ scripts/Token.js
@@ -12,7 +12,9 @@ isEmpty
 "use strict";
 
 import { ACTIVE_EFFECT_ICON } from "./const.js";
+import { tokenBounds } from "./util.js";
 import { UserCloneTargets } from "./UserCloneTargets.js";
+import { Settings } from "./settings.js";
 
 export const PATCHES = {};
 PATCHES.BASIC = {};
@@ -316,14 +318,16 @@ function _onDragLeftMove(wrapped, event) {
   }
 }
 
-function _onDragLeftDrop(wrapped, event) {
-  wrapped(event);
+async function _onDragLeftDrop(wrapped, event) {
+  const res = await wrapped(event);
 
   // Trigger each attached template to drag.
+  if ( !res || !event.interactionData.clones ) return res;
   for ( const clone of event.interactionData.clones ) {
     const attachedTemplates = clone.attachedTemplates;
-    for ( const template of attachedTemplates ) template._onDragLeftDrop(event);
+    for ( const template of attachedTemplates ) await template._onDragLeftDrop(event);
   }
+  return res;
 }
 
 function _onDragLeftCancel(wrapped, event) {
@@ -334,6 +338,47 @@ function _onDragLeftCancel(wrapped, event) {
     const attachedTemplates = clone.attachedTemplates;
     for ( const template of attachedTemplates ) template._onDragLeftCancel(event);
   }
+}
+
+/**
+ * Wrap Token.prototype._onHoverIn
+ * If the token is visible to the user and show-on-hover is enabled,
+ * display any templates that would otherwise be hidden.
+ */
+function _onHoverIn(wrapped, event, options) {
+  showHideTemplates(this, true);
+  return wrapped(event, options);
+}
+
+/**
+ * Wrap Token.prototype._onHoverOut
+ * If the token is visible to the user and show-on-hover is enabled,
+ * display any templates that would otherwise be hidden.
+ */
+function _onHoverOut(wrapped, event, options) {
+  showHideTemplates(this, false);
+  return wrapped(event, options);
+}
+
+/**
+ * Test every template on the scene.
+ * If the token shape overlaps a template, show it. Otherwise hide it.
+ * Skip if templates are not hidden, token is not visible, or show on hover not enabled.
+ */
+function showHideTemplates(token, show = true) {
+  const HIDE = Settings.KEYS.HIDE;
+  if ( !Settings.get(HIDE.HIGHLIGHTING)
+    || !Settings.get(HIDE.SHOW_ON_HOVER)
+    || !token.isVisible ) return;
+
+  const tBounds = tokenBounds(token);
+  canvas.templates.placeables.forEach(t => {
+    // Show or hide the highlight layer. See refreshMeasuredTemplate hook in MeasuredTemplate.js.
+    show &&= t.boundsOverlap(tBounds);
+    const alpha = show ? (t.document.hidden ? 0.5 : 1) : 0;
+    const hl = canvas.grid.getHighlightLayer(t.highlightId);
+    hl.alpha = alpha;
+  });
 }
 
 /**
@@ -363,4 +408,9 @@ PATCHES.BASIC.WRAPS = {
   _onDragLeftMove,
   _onDragLeftDrop,
   _onDragLeftCancel,
-  _refreshTarget };
+  _refreshTarget,
+
+  // Show on hover
+  _onHoverIn,
+  _onHoverOut
+};